requirements.txt
django
django-jwt-extended
djangorestframework
django-cors-headers